+++ src/main/java/com/dudar/runner/Runner.java
@@ -1,159 +1,93 @@
 package com.dudar.runner;
 
-import com.codeborne.selenide.WebDriverRunner;
-import com.dudar.InstaActor;
+import com.dudar.ActorsManager;
+import com.dudar.InstaActor2;
 import com.dudar.utils.Utilities;
+import com.dudar.utils.services.ActorActions;
 import com.dudar.utils.services.EmailService;
-import com.google.common.base.Strings;
-import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.log4j.Logger;
-import org.jetbrains.annotations.NotNull;
-import org.openqa.selenium.UnexpectedAlertBehaviour;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.remote.CapabilityType;
-import org.openqa.selenium.remote.RemoteWebDriver;
 
 import javax.mail.MessagingException;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.List;
-import java.util.Properties;
-
-import static com.codeborne.selenide.Selenide.*;
+import java.util.AbstractMap;
+import java.util.Date;
+import java.util.concurrent.TimeUnit;
 
 public class Runner {
 
     final static Logger logger = Logger.getLogger(Runner.class);
 
-    private static RemoteWebDriver driver;
-    private static Properties prop = new Properties();
+    public static void main(String[] args) throws InterruptedException, MessagingException {
 
-    public static void main(String[] args){
-        boolean debug;
-        List<String> hashTags;
-        String confFilePath = args[0];
-        try (InputStream input = new FileInputStream(confFilePath)) {
-            prop.load(input);
-        } catch (IOException ex) {
-            System.out.println("!!!ERROR on properties initialization");
-            ex.printStackTrace();
-            System.exit(1);
-        }
-        debug = Boolean.parseBoolean(prop.getProperty("debug.mode"));
-        //TODO better way to combine tags
-        hashTags = Utilities.getAllTags(args[1]);
-        hashTags.forEach(System.out::println);
-        System.out.println("Starting...");
-        List<String> currentTags = hashTags;
-        InstaActor actor = new InstaActor();
-        actor.setLogin(prop.getProperty("acc.user"))
-                .setPassword(prop.getProperty("acc.password"))
-                .enableLikes(prop.getProperty("likes.enabled"))
-                .enableComments(prop.getProperty("comments.enabled"))
-                .setMaxPostsCount(Integer.valueOf(prop.getProperty("posts.count")))
-                .setMinViewDelay(Integer.valueOf(prop.getProperty("view.min.delay")))
-                .setMaxViewDelay(Integer.valueOf(prop.getProperty("view.max.delay")))
-                .setMinVideoDelay(Integer.valueOf(prop.getProperty("video.min.delay")))
-                .setMaxVideoDelay(Integer.valueOf(prop.getProperty("video.max.delay")))
-                .setLikesPercentage(Integer.valueOf(prop.getProperty("likes.percentage")))
-                .setCommentsPercentage(Integer.valueOf(prop.getProperty("comments.percentage")));
-        actor.viewCurrentParameters();
-        try {
-            EmailService.generateAndSendEmail("Work Started</br>");
-        }
-        catch (MessagingException e) {
-            logger.debug(e.getLocalizedMessage());
-        }
-        while(actor.getCompletedTags().size() != hashTags.size())
+//        Map<String, InstaActor2> actors = new HashMap<>();
+//        String mainActorName = "InstaActor";
+//        actors.put(mainActorName, new InstaActor2(mainActorName));
+//
+//        actors.get(mainActorName).start();
+//
+        logger.info("Starting...");
+
+//        ActorsManager actors = ActorsManager.getInstance();
+
+        Date lastActionDate = new Date();
+
+        AbstractMap.SimpleEntry<String, ActorActions> currentAction;
+
+        EmailService.generateAndSendEmail("InstaActor service is UP and running");
+
+        while(true)
         {
-            try {
-                initDriver(debug);
-                actor
-                        .loadTags(currentTags)
-                        .start();
+//            currentAction = new AbstractMap.SimpleEntry<>("inline", ActorActions.START);
+            currentAction = EmailService.getActionFromEmail(Utilities.getActionsUserEmail(), lastActionDate);
+            lastActionDate = new Date();
+
+            if(currentAction.getValue() == ActorActions.STATUS
+            & currentAction.getKey().equals("ALL")){
+                String message = "<p>Registered actors:";
+                message += "<p>" + ActorsManager.getInstance().getAllStatus();
+
+                EmailService.generateAndSendEmail(message);
             }
-            catch (AssertionError err) {
-                currentTags = exceptionClose(hashTags, actor, "ASSERT ERRROR\n"+err.getLocalizedMessage());
+
+            if(currentAction.getValue() == ActorActions.ABORT){
+                EmailService.generateAndSendEmail("<h1>!!!STOP FOR EXECUTION!!!");
+                logger.info("END");
+                System.exit(0);
             }
-            catch (Exception ex){
-                System.out.println("!!!OTHER EXCEPTION");
-                currentTags = exceptionClose(hashTags, actor, "EXCEPTION\n"+ex.getLocalizedMessage());
+
+            if((currentAction.getValue() != ActorActions.UNDEFINED)
+                    & !StringUtils.isEmpty(currentAction.getKey())){
+
+                //DO SOME STUFF
+                logger.info(currentAction.getKey() + " --- " + currentAction.getValue());
+
+                ActorsManager.getInstance().proceedAction(currentAction);
+
+
             }
+
+            //
+            TimeUnit.SECONDS.sleep(60);
         }
-        clearBrowserLocalStorage();
-        clearBrowserCookies();
-        WebDriverRunner.getWebDriver().quit();
-        getCurrentStateForCompletedActions(actor);
+
+    }
+
+    private static boolean statusRequestedEmailService(InstaActor2 actor){
         try {
-            EmailService.generateAndSendEmail("Work Completed</br> + Likes added: " + actor.getTotalLikes() + "</br>");
+            return EmailService.isStatusRequestMessage(Utilities.getActionsUserEmail(), actor.getName(), actor.getCreationDate());
+        } catch (MessagingException e) {
+            e.printStackTrace();
         }
-        catch (MessagingException e) {
-            logger.debug(e.getLocalizedMessage());
-        }
-        System.out.println("Shutting down!");
+        return false;
     }
 
-    @NotNull
-    private static List<String> exceptionClose(List<String> hashTags, InstaActor actor, String issueMessage) {
-        List<String> currentTags;
-        logger.error("UNEXPECTED STOP INFO:\n");
-        logger.error(issueMessage);
-
-        getCurrentStateForCompletedActions(actor);
-        currentTags = (List<String>) CollectionUtils.disjunction(hashTags, actor.getCompletedTags());
-        try{
-            clearBrowserLocalStorage();
-            clearBrowserCookies();
-            WebDriverRunner.getWebDriver().quit();
+    private static boolean shouldStopEmailService(InstaActor2 actor){
+        try {
+            return EmailService.isStopMessage(Utilities.getActionsUserEmail(), actor.getName(), actor.getCreationDate());
+        } catch (MessagingException e) {
+            e.printStackTrace();
         }
-        catch (Exception ex){
-            logger.error("!!!Can't terminate driver");
-            logger.error(ex.getLocalizedMessage());
-        }
-        sleep(5000);
-        return currentTags;
+        return false;
     }
 
-    private static void getCurrentStateForCompletedActions(InstaActor actor) {
-        System.out.println("Likes added: " + actor.getTotalLikes());
-        System.out.println("Comments added: " + actor.getTotalComments());
-        System.out.println("Completed tags:");
-        actor.getCompletedTags().forEach(System.out::println);
-        System.out.println("Defected tags");
-        actor.printDefectedTags();
-        System.out.println("**********************InstaActor*******************");
-    }
-
-    private static void initDriver(boolean debug) {
-        if(!debug) {
-            String seleniumHub = System.getenv("HUB_HOST");
-            String seleniumHubPort = System.getenv("HUB_PORT");
-            if(Strings.isNullOrEmpty(seleniumHub) || Strings.isNullOrEmpty(seleniumHubPort)){
-                seleniumHub = prop.getProperty("hub.host");
-                seleniumHubPort = prop.getProperty("hub.port");
-            }
-            String gridHubUrl = "http://" + seleniumHub + ":" + seleniumHubPort;
-            //Check grid status
-            Utilities.checkGridStatus(gridHubUrl);
-            try {
-                ChromeOptions chromeOptions = new ChromeOptions();
-                chromeOptions.setCapability(CapabilityType.UNEXPECTED_ALERT_BEHAVIOUR,
-                        UnexpectedAlertBehaviour.IGNORE);
-                chromeOptions.setHeadless(true);
-                driver = new RemoteWebDriver(new URL(gridHubUrl+"/wd/hub"), chromeOptions);
-            } catch (MalformedURLException e) {
-                System.out.println("!!!Can't init DRIVER");
-                System.out.println("Error message: " + e.getLocalizedMessage());
-                driver = null;
-            }
-        }
-        else {
-            driver = new ChromeDriver();
-        }
-        WebDriverRunner.setWebDriver(driver);
-    }
 }
